+++ CHANGELOG.md
@@ -64,4 +64,6 @@ Version 0.0.11 (TBD)
 * Add lt_ operator, and default implementations of le_ and or_
 * Improve error when an operator overload fails.
 * add min_window_period to rolling_average
+* Fix a bug in reduce over a TSL
+